+++ Morpheus.Tests/YoutubeFeedServiceTests.cs
@@ -43,6 +43,43 @@ public class YoutubeFeedServiceTests
     }
 
     [Fact]
+    public void ParseEntries_PrefersAlternateLinkOverFeedMetadataLinks()
+    {
+        XDocument document = XDocument.Parse("""
+            <feed xmlns="http://www.w3.org/2005/Atom"
+                  xmlns:yt="http://www.youtube.com/xml/schemas/2015">
+              <entry>
+                <yt:videoId>video-id</yt:videoId>
+                <link rel="self" href="https://www.youtube.com/feeds/videos.xml?video_id=video-id" />
+                <link rel="alternate" href="https://www.youtube.com/watch?v=video-id" />
+              </entry>
+            </feed>
+            """);
+
+        YoutubeFeedService.VideoEntry entry = Assert.Single(YoutubeFeedService.ParseEntries(document));
+        Assert.Equal("https://www.youtube.com/watch?v=video-id", entry.Link);
+    }
+
+    [Theory]
+    [InlineData("<link rel=\"self\" href=\"https://www.youtube.com/feeds/videos.xml?video_id=video-id\" />")]
+    [InlineData("<link rel=\"alternate\" href=\"  \" /><link rel=\"self\" href=\"https://www.youtube.com/feeds/videos.xml?video_id=video-id\" />")]
+    public void ParseEntries_WhenOnlyMetadataOrBlankAlternateLinksExist_UsesCanonicalVideoUrl(string linkElements)
+    {
+        XDocument document = XDocument.Parse($"""
+            <feed xmlns="http://www.w3.org/2005/Atom"
+                  xmlns:yt="http://www.youtube.com/xml/schemas/2015">
+              <entry>
+                <yt:videoId>video-id</yt:videoId>
+                {linkElements}
+              </entry>
+            </feed>
+            """);
+
+        YoutubeFeedService.VideoEntry entry = Assert.Single(YoutubeFeedService.ParseEntries(document));
+        Assert.Equal("https://www.youtube.com/watch?v=video-id", entry.Link);
+    }
+
+    [Fact]
     public void ParseEntries_TrimsWhitespaceAroundVideoIds()
     {
         XDocument document = XDocument.Parse("""